requirements.txt
requests
pyqt6>=6.7.1
websocket-client>=1.0.0
